Create separate files for each category. Generalised test construction and deconstruction.
